파일 업로드 위치를 aws s3로 변경 - 기존에는 서버에다가 저장했다면, 이제는 cloud에 저장

+++ teamEmpty/src/main/java/com/itbank/simpleboard/service/LectureService.java
@@ -29,24 +29,26 @@ public class LectureService {
     }
 
     public List<LectureDto> selectAll(String searchType, String keyword) {
-        return lectureRepository.getLectureListDtos(searchType,keyword);
+        return lectureRepository.getLectureListDtos(searchType, keyword);
     }
 
 
     @Transactional
     public String planUpload(MultipartFile plan, Long lectureIdx) {
         String stringPlan = null;
-        Optional<Lecture> optionalLecture = lectureRepository.findById(lectureIdx);
-        if (optionalLecture.isPresent()) {
-            Lecture lecture = optionalLecture.get();
-            if (lecture.getPlan() != null) {
-                fileComponent.deleteFile(lecture.getPlan(), "syllabus");
-            }
-            String syllabus = fileComponent.upload(plan, "syllabus");
-            if (syllabus != null) {
-                lecture.setPlan(syllabus);
-                lectureRepository.save(lecture);
-                stringPlan = syllabus;
+        if (!plan.isEmpty()) {
+            Optional<Lecture> optionalLecture = lectureRepository.findById(lectureIdx);
+            if (optionalLecture.isPresent()) {
+                Lecture lecture = optionalLecture.get();
+                if (lecture.getPlan() != null) {
+                    fileComponent.deleteFile(lecture.getPlan(), "syllabus");
+                }
+                String syllabus = fileComponent.upload(plan, "syllabus");
+                if (syllabus != null) {
+                    lecture.setPlan(syllabus);
+                    lectureRepository.save(lecture);
+                    stringPlan = syllabus;
+                }
             }
         }
         return stringPlan;
